requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
jieba3k==0.35.1
lxml==4.2.3
newspaper3k==0.2.6
nltk==3.3
Pillow==5.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.23
